requirements.txt
scapy==2.6.1
requests>=2.30.0
python-nmap>=0.7.1